Fix issue with task delegation and add missing `name` parameter to generated delegation Task
The `name` parameter had to be added to fix the following issue:
```
I encountered an error while trying to use the tool. This was the error: 1 validation error for Task
name
  Field required [type=missing, input_value={'description': 'conduct ...or the context shared.'}, input_type=dict]
    For further information visit https://errors.pydantic.dev/2.7/v/missing.
 Tool Delegate work to coworker accepts these inputs: Delegate a specific task to one of the following coworkers: web researcher
The input to this tool should be the coworker, the task you want them to do, and ALL necessary context to execute the task, they know nothing about the task, so share absolute everything you know, don't reference things but instead explain them.
```